requirements.txt
Django==1.5.5
Mezzanine==1.4.16
Pillow==2.2.1
South==0.8.4
astroid==1.0.1
bleach==1.2.2
coverage==3.6
dj-database-url==0.2.2
dj-static==0.0.5
django-appconf==0.6
django-compressor==1.3
django-extensions==1.2.5
django-less==0.7.2
django-redis-sessions==0.4.0
django-rosetta==0.7.2
django-toolbelt==0.0.1
filebrowser-safe==0.2.31
grappelli-safe==0.2.23
gunicorn==18.0
html5lib==0.95
logilab-common==0.60.0
oauthlib==0.6.0
pep8==1.4.6
pyflakes==0.7.3
pytz==2013.8
redis==2.8.0
requests==1.2.3
requests-oauthlib==0.3.3
six==1.4.1
static==0.4
transifex-client==0.10
travis==0.0.1
wsgiref==0.1.2
xhtml2pdf==0.0.5
django-mandrill==0.1.0
django-endless-pagination==2.0
django-gmapi==1.0.1